Create core package
